refactoring project structure: moved helpers files
