+++ lib/payloads/credentialsPayload.js
@@ -1,5 +1,5 @@
 const { APP_ID } = require('../../src/data/constants');
-const { timestamp, nonce } = require('../helpers');
+const { timestamp, nonce } = require('../../src/helpers/utilities');
 
 const credentialsPayload = ({ email, password }) => ({
   appid: APP_ID,
