+++ lib/payloads/wssUpdatePayload.js
@@ -1,4 +1,4 @@
-const { timestamp } = require('../helpers');
+const { timestamp } = require('../../src/helpers/utilities');
 
 const wssUpdatePayload = ({ apiKey, deviceId, params }) => {
   const payload = {
