+++ lib/payloads/zeroConfUpdatePayload.js
@@ -1,5 +1,5 @@
 const { encryptationData } = require('../../src/helpers/ewelink');
-const { timestamp } = require('../helpers');
+const { timestamp } = require('../../src/helpers/utilities');
 
 const zeroConfUpdatePayload = (selfApikey, deviceId, deviceKey, params) => {
   const encryptedData = encryptationData(JSON.stringify(params), deviceKey);
