+++ src/classes/ChangeStateZeroconf.js
@@ -2,7 +2,7 @@ const rp = require('request-promise');
 
 const WebSocket = require('./WebSocket');
 const payloads = require('../../lib/payloads');
-const { _get } = require('../../lib/helpers');
+const { _get } = require('../helpers/utilities');
 
 class ChangeStateZeroconf extends WebSocket {
   static async set({ url, device, params, switches, state }) {
